feat: add Copyright component and integrate into Blog, Calendar, and Image components
