+++ src/lib/components/ui/Blog.svelte
@@ -45,7 +45,6 @@
 											enabled: post.copyright.enabled,
 											name: post.copyright.name,
 											url: post.copyright.url || '',
-											compact: true,
 											size: 'xs'
 										}
 									]
@@ -99,7 +98,8 @@
 					</div>
 
 					<div class="mt-1 text-sm text-zinc-600 dark:text-zinc-400">
-						Der Blog konnte gerade nicht geladen werden. Bitte versuche es in ein paar Minuten erneut oder lade die Seite neu.
+						Der Blog konnte gerade nicht geladen werden. Bitte versuche es in ein paar Minuten
+						erneut oder lade die Seite neu.
 					</div>
 
 					<!-- Actions -->
